backend/requirements.txt
alembic==1.16.5
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
async-timeout==5.0.1
bcrypt==3.2.2
cffi==2.0.0
click==8.1.8
colorama==0.4.6
dnspython==2.7.0
email-validator==2.3.0
exceptiongroup==1.3.1
fastapi==0.128.8
greenlet==3.2.5
h11==0.16.0
httptools==0.7.1
idna==3.11
mako==1.3.10
markupsafe==3.0.3
passlib==1.7.4
psycopg==3.2.13
psycopg-binary==3.2.13
pycparser==2.23
pydantic==2.12.5
pydantic-core==2.41.5
pyjwt==2.12.1
python-dotenv==1.2.1
pyyaml==6.0.3
redis==7.0.1
sqlalchemy==2.0.48
starlette==0.49.3
tomli==2.4.0
typing-extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.3
uvicorn==0.39.0
watchfiles==1.1.1
websockets==15.0.1
